requirements.txt
numpy==1.21.5
PyYAML==6.0
torch==1.12.1
matplotlib
